+++ api/src/functionalTest/java/uk/gov/hmcts/reform/divorce/caseapi/ServiceAuthValidationTest.java
@@ -1,0 +1,68 @@
+package uk.gov.hmcts.reform.divorce.caseapi;
+
+import io.restassured.RestAssured;
+import io.restassured.response.Response;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import uk.gov.hmcts.reform.divorce.caseapi.model.CaseDetails;
+import uk.gov.hmcts.reform.divorce.caseapi.model.CcdCallbackRequest;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
+import static org.springframework.http.HttpStatus.FORBIDDEN;
+import static org.springframework.http.HttpStatus.UNAUTHORIZED;
+import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
+import static uk.gov.hmcts.reform.divorce.caseapi.enums.NotificationConstants.SERVICE_AUTHORIZATION;
+
+@SpringBootTest
+public class ServiceAuthValidationTest extends FunctionalTestSuite {
+    @Test
+    public void shouldReturn401WhenServiceAuthTokenIsInvalid() {
+        Response response = RestAssured
+            .given()
+            .relaxedHTTPSValidation()
+            .baseUri(testUrl)
+            .header(CONTENT_TYPE, APPLICATION_JSON_VALUE)
+            .header(SERVICE_AUTHORIZATION, "Bearer invalid_token")
+            .body(
+                CcdCallbackRequest
+                    .builder()
+                    .caseDetails(
+                        CaseDetails
+                            .builder()
+                            .caseData(caseData())
+                            .build()
+                    )
+                    .build()
+            )
+            .when()
+            .post(SAVE_AND_SIGN_OUT_CALLBACK_URL);
+
+        assertThat(response.getStatusCode()).isEqualTo(UNAUTHORIZED.value());
+    }
+
+    @Test
+    public void shouldReturn403WhenServiceIsNotAuthorised() {
+        Response response = RestAssured
+            .given()
+            .relaxedHTTPSValidation()
+            .baseUri(testUrl)
+            .header(CONTENT_TYPE, APPLICATION_JSON_VALUE)
+            .header(SERVICE_AUTHORIZATION, generateServiceAuthTokenFor("ccd_gw"))
+            .body(
+                CcdCallbackRequest
+                    .builder()
+                    .caseDetails(
+                        CaseDetails
+                            .builder()
+                            .caseData(caseData())
+                            .build()
+                    )
+                    .build()
+            )
+            .when()
+            .post(SAVE_AND_SIGN_OUT_CALLBACK_URL);
+
+        assertThat(response.getStatusCode()).isEqualTo(FORBIDDEN.value());
+    }
+}
