+++ api/src/integrationTest/java/uk/gov/hmcts/reform/divorce/caseapi/controllers/SaveAndCloseControllerTest.java
@@ -9,7 +9,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
-import uk.gov.hmcts.reform.authorisation.validators.AuthTokenValidator;
+import uk.gov.hmcts.reform.divorce.caseapi.config.WebMvcConfig;
+import uk.gov.hmcts.reform.divorce.caseapi.config.interceptors.RequestInterceptor;
 import uk.gov.hmcts.reform.divorce.caseapi.exceptions.NotificationException;
 import uk.gov.hmcts.reform.divorce.caseapi.notification.handler.SaveAndSignOutNotificationHandler;
 import uk.gov.hmcts.reform.divorce.caseapi.service.NotificationService;
@@ -20,14 +21,12 @@ import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
-import static org.mockito.Mockito.when;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static uk.gov.hmcts.reform.divorce.caseapi.TestConstants.API_URL;
 import static uk.gov.hmcts.reform.divorce.caseapi.TestConstants.AUTH_HEADER_VALUE;
-import static uk.gov.hmcts.reform.divorce.caseapi.TestConstants.CCD_DATA;
 import static uk.gov.hmcts.reform.divorce.caseapi.TestConstants.SERVICE_AUTHORIZATION;
 import static uk.gov.hmcts.reform.divorce.caseapi.TestConstants.TEST_USER_EMAIL;
 import static uk.gov.hmcts.reform.divorce.caseapi.caseapi.util.TestDataHelper.callbackRequest;
@@ -42,23 +41,24 @@ public class SaveAndCloseControllerTest {
     @Autowired
     private MockMvc mockMvc;
 
-    @MockBean
-    private NotificationService notificationService;
-
     @Autowired
     @SuppressWarnings("PMD.UnusedPrivateField")
     private SaveAndSignOutNotificationHandler saveAndSignOutNotificationHandler;
 
-    @MockBean
-    private AuthTokenValidator validator;
-
     @Autowired
     private ObjectMapper objectMapper;
 
+    @MockBean
+    private NotificationService notificationService;
+
+    @MockBean
+    private RequestInterceptor requestInterceptor;
+
+    @MockBean
+    private WebMvcConfig webMvcConfig;
+
     @Test
     public void givenValidCaseDataWhenCallbackIsInvokedThenSendEmail() throws Exception {
-        when(validator.getServiceName(AUTH_HEADER_VALUE)).thenReturn(CCD_DATA);
-
         mockMvc.perform(post(API_URL)
             .contentType(APPLICATION_JSON)
             .header(SERVICE_AUTHORIZATION, AUTH_HEADER_VALUE)
@@ -68,9 +68,8 @@ public class SaveAndCloseControllerTest {
 
         verify(notificationService)
             .sendEmail(eq(TEST_USER_EMAIL), eq("70dd0a1e-047f-4baa-993a-e722db17d8ac"), anyMap(), eq(ENGLISH));
-        verify(validator).getServiceName(AUTH_HEADER_VALUE);
 
-        verifyNoMoreInteractions(notificationService, validator);
+        verifyNoMoreInteractions(notificationService);
     }
 
     @Test
@@ -84,8 +83,6 @@ public class SaveAndCloseControllerTest {
 
     @Test
     public void givenSendEmailThrowsExceptionWhenCallbackIsInvokedThenReturnBadRequest() throws Exception {
-        when(validator.getServiceName(AUTH_HEADER_VALUE)).thenReturn(CCD_DATA);
-
         doThrow(new NotificationException(new NotificationClientException("All template params not passed")))
             .when(notificationService).sendEmail(
             eq(TEST_USER_EMAIL),
@@ -100,7 +97,5 @@ public class SaveAndCloseControllerTest {
             .accept(APPLICATION_JSON))
             .andExpect(status().isBadRequest())
             .andExpect(content().string("All template params not passed"));
-
-        verify(validator).getServiceName(AUTH_HEADER_VALUE);
     }
 }
